feature/93 - moves runway tests to runway dir and updates imports
