+++ test/airport/runway/RunwayModel.spec.js
@@ -1,8 +1,8 @@
 import ava from 'ava';
 
-import RunwayModel from '../../src/assets/scripts/client/airport/runway/RunwayModel';
-import { airportModelFixture } from '../fixtures/airportFixtures';
-import { AIRPORT_JSON_KLAS_MOCK } from './_mocks/airportJsonMock';
+import RunwayModel from '../../../src/assets/scripts/client/airport/runway/RunwayModel';
+import { airportModelFixture } from '../../fixtures/airportFixtures';
+import { AIRPORT_JSON_KLAS_MOCK } from '../_mocks/airportJsonMock';
 
 // taken from KLAS
 // "runways":[
